connect players to player profile
