+++ src/app/shared/interfaces/dev-goal.ts
@@ -1,0 +1,5 @@
+export interface DevelopmentGoal {
+  id: number;
+  title: string;
+  completed: boolean;
+}
