+++ src/app/shared/interfaces/dev-rating.model.ts
@@ -1,0 +1,10 @@
+export interface DevelopmentRating {
+  category: DevelopmentCategory;
+  value: number;
+}
+
+type DevelopmentCategory =
+  | 'Technical'
+  | 'Physical'
+  | 'Tactical'
+  | 'Teamwork';
