+++ src/app/shared/interfaces/journey-event.model.ts
@@ -1,0 +1,6 @@
+export interface JourneyEvent {
+  id: number;
+  year: number;
+  title: string;
+  description?: string;
+}
