+++ src/app/shared/interfaces/personal-best.model.ts
@@ -1,0 +1,6 @@
+export interface PersonalBest {
+  id: number;
+  label: string;
+  value: string;
+  previousValue?: string;
+}
