+++ src/app/shared/interfaces/player-badge.model.ts
@@ -1,0 +1,7 @@
+export interface PlayerBadge {
+  id: number;
+  name: string;
+  icon: string;
+  description: string;
+  earnedDate: string;
+}
