+++ src/app/shared/interfaces/player-highlight.model.ts
@@ -1,0 +1,6 @@
+export interface Highlight {
+  id: number;
+  title: string;
+  thumbnail: string;
+  videoUrl?: string;
+}
