+++ src/app/shared/interfaces/player-stat.model.ts
@@ -1,0 +1,4 @@
+export interface PlayerStat {
+  label: string;
+  value: string | number;
+}
