+++ src/styles.scss
@@ -57,3 +57,14 @@ p, ul, strong {
     display: none;
   }
 }
+
+
+.cdk-overlay-backdrop.cdk-overlay-dark-backdrop {
+  background: rgba(4, 14, 9, 0.78);
+}
+
+.cdk-overlay-backdrop.cdk-overlay-dark-backdrop {
+  background: rgba(4, 14, 9, 0.78);
+
+  backdrop-filter: blur(4px);
+}
